+++ python-threatexchange/threatexchange/clients/ncmec/tests/test_hash_api.py
@@ -3,7 +3,7 @@ from tkinter import Image
 from unittest.mock import Mock
 import pytest
 import requests
-from threatexchange.ncmec.hash_api import (
+from threatexchange.clients.ncmec.hash_api import (
     NCMECEndpoint,
     NCMECEntryType,
     NCMECEntryUpdate,
